+++ src/openpile/soilmodels.py
@@ -1165,7 +1165,7 @@ class API_sand(LateralModel):
 
     See also
     --------
-    :py:func:`openpile.utils.py_curves.api_sand`, :py:func:`openpile.utils.multipliers.durkhop`
+    :py:func:`openpile.utils.py_curves.api_sand`, :py:func:`openpile.utils.hooks.durkhop`
 
     """